Refactor CanvasCommand
