+++ canvas/__main__.py
@@ -4,14 +4,13 @@ import os
 from canvasapi import Canvas
 from dotenv import load_dotenv
 
-from .cli.factory import CommandFactory
-from .cli.parse import get_parser
+from canvas.cli.manager import CommandManager
 
 
 def main() -> None:
     """The main entry point for the CLI."""
 
-    parser = get_parser()
+    parser = CommandManager.get_parser()
     args = parser.parse_args()
 
     # Print help if no arguments are passed.
@@ -21,11 +20,10 @@ def main() -> None:
 
     API_URL = os.getenv("API_URL")
     API_KEY = os.getenv("API_KEY")
-
     client = Canvas(API_URL, API_KEY)
 
     # Run the command
-    cmd = CommandFactory.from_args(
+    cmd = CommandManager.from_args(
         args, client  # pyright: ignore reportArgumentType
     )
     cmd.execute()
