+++ canvas/cli/__init__.py
@@ -1,6 +1,4 @@
 from __future__ import annotations
 
 # flake8: noqa
-from .parse import *
-from .factory import *
 from .base import *
