+++ canvas/cli/commands/__init__.py
@@ -2,3 +2,5 @@ from __future__ import annotations
 
 # flake8: noqa
 from .init import *
+from .add import *
+from .status import *
