+++ tests/command/test_command.py
@@ -15,7 +15,7 @@ from canvasapi.user import User
 
 from canvas.cli.base import NotCanvasCourseException
 from canvas.cli.commands.init import InitCommand
-from canvas.cli.factory import CommandFactory, CommandNotFoundException
+from canvas.cli.manager import CommandManager, CommandNotFoundException
 
 
 @pytest.fixture
@@ -39,7 +39,7 @@ def test_from_args_init(mock_client: Canvas) -> None:
     init_args = Namespace(command="init", course_id="1")
 
     assert isinstance(
-        CommandFactory.from_args(init_args, mock_client),
+        CommandManager.from_args(init_args, mock_client),
         InitCommand,
     )
 
@@ -49,7 +49,7 @@ def test_from_args_fail(mock_client: Canvas) -> None:
     invalid_args = Namespace(command="fake-command")
 
     with pytest.raises(CommandNotFoundException):
-        CommandFactory.from_args(invalid_args, mock_client)
+        CommandManager.from_args(invalid_args, mock_client)
 
 
 def test_find_course_root_success(init_command: InitCommand) -> None:
@@ -64,7 +64,7 @@ def test_find_course_root_success(init_command: InitCommand) -> None:
     os.chdir(temp_dir)
 
     # Find course root
-    assert init_command.find_course_root() == temp_dir
+    assert init_command.get_course_root() == temp_dir
 
     # Cleanup fake course
     os.chdir(file_dir)
@@ -74,4 +74,4 @@ def test_find_course_root_success(init_command: InitCommand) -> None:
 def test_find_course_root_fail(init_command: InitCommand) -> None:
     """Test finding the course root when not in a course."""
     with pytest.raises(NotCanvasCourseException):
-        init_command.find_course_root()
+        init_command.get_course_root()
